+++ libraries/DS3232/DS3232.h
@@ -3,7 +3,7 @@
 //    FILE: DS3232.h
 //  AUTHOR: Rob Tillaart
 // PURPOSE: Arduino library for DS3232 RTC (minimalistic)
-// VERSION: 0.4.0
+// VERSION: 0.4.1
 //    DATE: 2011-01-21
 //     URL: https://github.com/RobTillaart/DS3232
 
@@ -12,7 +12,7 @@
 #include "Wire.h"
 
 
-#define DS3232_LIB_VERSION         "0.4.0"
+#define DS3232_LIB_VERSION         "0.4.1"
 
 
 //  ERROR CODES
@@ -33,22 +33,24 @@
 #define DS3232_SRAM_BASE            0x14
 
 
-class DS3232
+class DS3231
 {
 public:
-  //  CONSTRUCTOR
-  DS3232(TwoWire *wire = &Wire);
+  //  Constructor
+  DS3231(TwoWire *wire = &Wire);
   int      begin();
   bool     isConnected();
   uint8_t  getAddress();
   uint16_t getType();
+
 
   //  BASE RTC
   int      read();
   int      write();
   uint32_t lastRead();
 
-  //  GETTERS
+
+  //  Getters
   uint8_t  seconds();
   uint8_t  minutes();
   uint8_t  hours();
@@ -57,7 +59,8 @@ public:
   uint8_t  month();
   uint8_t  year();
 
-  //  SETTERS
+
+  //  Setters
   void     setSeconds(uint8_t value);
   void     setMinutes(uint8_t value);
   void     setHours(uint8_t value);
@@ -66,6 +69,9 @@ public:
   void     setMonth(uint8_t value);
   void     setYear(uint8_t value);
 
+  //  Temperature
+  float    getTemperature();
+
 
   //  LOW LEVEL access to all registers
   //  check datasheet for details of registers.
@@ -73,7 +79,7 @@ public:
   int      readRegister(uint8_t reg);
   int      writeRegister(uint8_t reg, uint8_t value);
 
-  //  DEBUG
+  //  Debug
   int      lastRv() { return _rv; };
 
 
@@ -86,7 +92,7 @@ protected:
   uint32_t  _lastRead = 0;
 
   int       _rv;
-  uint16_t  _type = 3232;
+  uint16_t  _type = 3231;
 
   uint8_t   dec2bcd(uint8_t value);
   uint8_t   bcd2dec(uint8_t value);
@@ -97,10 +103,27 @@ protected:
 //
 //  DERIVED CLASSES
 //
-class DS3231 : public DS3232
+class DS3232 : public DS3231
 {
 public:
-  DS3231(TwoWire *wire = &Wire);
+  DS3232(TwoWire *wire = &Wire);
+
+  //  EXPERIMENTAL SRAM SUPPORT
+  //
+  //  SRAM 236 bytes, register 0x14-0xFF
+  //  index = 0..235 == 0x00..0xEB
+  //  236 bytes = 59 x uint32_t (e.g. time stamps)
+  //  note: no boundary check
+  //  TODO: optimize read/write multiple bytes at once.
+  int      SRAMwrite8(uint8_t index, uint8_t value);
+  int      SRAMwrite16(uint8_t index, uint16_t value);
+  int      SRAMwrite32(uint8_t index, uint32_t value);
+  uint8_t  SRAMread8(uint8_t index);
+  uint16_t SRAMread16(uint8_t index);
+  uint32_t SRAMread32(uint8_t index);
+
+private:
+
 };