+++ libraries/DS3232/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.4.1] - 2024-05-17
+- add **float getTemperature()**
+- make DS3231 the base class as the DS3232 has EEPROM (logical)
+- add experimental SRAM support 
+  - **SRAMwrite8/16/32()** + **SRAMread8/16/32()**
+- add example to test SRAM for DS3232.
+- update readme.md.
+- minor edits
+
 ## [0.4.0] - 2024-04-05
 - fix #2, add **setWeekDay()**
 - update examples